requirements.txt
youtube-transcript-api
google-generativeai
beautifulsoup4
markdown